Update resistance check
The resistance check now gathers 75 data points and calculates the
resistance using every data point against every other data point, as
long as neither the current nor voltage values are the same between two
points. This is similar to the handshake math problem.

The data points are gathered by putting the cell into 4 states of load
based on the binary values of the two loads. A pseudorandom process is
used to determine which state is selected in which order so as to avoid
a "drift" from the battery voltgae dropping during the resistance check.

The resistance check now also calculates the energy spent during the
resistance check. It is not completely accurate, as it only measures the
current at the end of an interval, but the incandescent bulb will pull
more current at the beginning of the interval, if the filaments start
cold.

The resistance check now returns values which vary by roughly 10%
between tests, a great improvement over previous test algorithms.